requirements.txt
streamlit
sxtwl
lunar_python
google-generativeai